+++ packages/network-controller/src/rpc-service/rpc-service.ts
@@ -63,9 +63,8 @@ export type RpcServiceOptions = {
    */
   policyOptions?: Omit<CreateServicePolicyOptions, 'retryFilterPolicy'>;
   /**
-   * A function that checks if the user is currently offline. If it returns true,
-   * connection errors will not be retried, preventing degraded and break
-   * callbacks from being triggered.
+   * A function that checks if the user is currently offline. If it returns
+   * true, onDegraded and onBreak callbacks will not be called.
    */
   isOffline: () => boolean;
 };
@@ -313,6 +312,12 @@ export class RpcService {
   readonly #fetchOptions: FetchOptions;
 
   /**
+   * A function that checks if the user is currently offline. If it returns
+   * true, onDegraded and onBreak callbacks will not be called.
+   */
+  readonly #isOffline: () => boolean;
+
+  /**
    * A `loglevel` logger.
    */
   readonly #logger: RpcServiceOptions['logger'];
@@ -347,18 +352,13 @@ export class RpcService {
     );
     this.endpointUrl = stripCredentialsFromUrl(normalizedUrl);
     this.#logger = logger;
+    this.#isOffline = isOffline;
 
     this.#policy = createServicePolicy({
       maxRetries: DEFAULT_MAX_RETRIES,
       maxConsecutiveFailures: DEFAULT_MAX_CONSECUTIVE_FAILURES,
       ...policyOptions,
       retryFilterPolicy: handleWhen((error) => {
-        // If user is offline, don't retry any errors
-        // This prevents degraded/break callbacks from being triggered
-        if (isOffline()) {
-          return false;
-        }
-
         return (
           // Ignore errors where the request failed to establish
           isConnectionError(error) ||
@@ -442,7 +442,7 @@ export class RpcService {
       // doesn't function the way it is intended, at least in the context of an
       // RpcService. However, we are making a bet that we won't need to use it
       // other than how we are already using it.
-      if (!('isolated' in data)) {
+      if (!hasProperty(data, 'isolated') && !this.#isOffline()) {
         listener({
           ...data,
           endpointUrl: this.endpointUrl.toString(),
@@ -471,7 +471,7 @@ export class RpcService {
           endpointUrl: this.endpointUrl.toString(),
           rpcMethodName: this.#currentRpcMethodName,
         });
-      } else {
+      } else if (!this.#isOffline()) {
         listener({
           ...data,
           endpointUrl: this.endpointUrl.toString(),